Finds similarity between sentences, but there are a couple issues: 1) When making the vocabulary set based on the two sentences, there are some words that are not found on Wordnet, and therefore do not yield a synset. The approach of having the vocabulary be a set of synsets is not good enough, since there are some words that are either in another language or just plain misspelled. Possible solution: Make the vocabulary two lists, a set of words (as it was before), and a set of their respective synset or None if no synset is found. Also, try running textblobs autocorrect on each word first, and keep the corrected word if it happens to be incorrect. So, in conclusion. First preprocess the sentence word by word by trying to correct it with textblob if possible, and then run the POS tagger, and then store the word on two lists (one with the word and one with the synset) 2) Links between sentences of very different lengths are not found by the current textblob semantic similarity approach. 3) Apparently semantics is not a proper way to identify links.
Proposed solution: The comment structure should be front and center when
trying to find links between sentences. Do a drill down approach. Use
gensim to find similarity between each sentence and all other comments
and the article (separate the article into groups of sentences in order,
maybe clusters can be done with lexrank).
Once I have the links between a sentence and comments or sets of
sentences in the article. Run the semantic similarity algorithm on every
pair of sentences in those clusters.

Step 1: Grouping
Option 1
First approach to grouping, make groups of sentences in order.

Option 2
Second approach to grouping, use lexrank to get the similarity matrix of all
sentences in the article, and all sentences in the comments. Cluster
sentences in article in and in comments. And use those clusters as
groups.

Step 2: Finding links between sentences and groups
1. Make a corpora with the groups of sentences found above. Each group is a doc.
2. Go through each sentence in the comments and use them as queries to get
a set of related docs to each sentence. A sentence is related to a doc
if its similarity is over a specific treshold. After this, I have a set
of document and sentence pairs.
3. Go through each document and sentence pair. For each of them, check
similarity between the sentence and each document sentence. This will id
the specific sentence within a document (which could be a cluster or a
paragraph depending of the grouping strategy done above) that a sentence
is semantically similar to. If the link is not strong enough to any
sentence (not above a specific tresh) then the sentence is semantically
similar to the document as a whole.

Step 3: Store the links in a dictionary form.